"removed extraneous folders from workspace; resolved servlet error; updated web.xml to add servlets; removed asome extraneous files; 'Unable to compile class for jsp' error stil present - unable to resolve custom class and test filter until fixed"
