+++ .vscode/settings.json
@@ -1,3 +1,7 @@
 {
-    "java.configuration.updateBuildConfiguration": "interactive"
+    "java.configuration.updateBuildConfiguration": "automatic",
+    "maven.view": "flat",
+    "maven.pomfile.autoUpdateEffectivePOM": true,
+    "maven.terminal.useJavaHome": true,
+    "maven.executable.path": "C:\\apache-maven-3.8.3\\bin"
 }
