+++ EcoMap/pom.xml
@@ -27,6 +27,13 @@
     </dependencyManagement>
 
   <dependencies>
+
+    <dependency>
+      <groupId>org.glassfish</groupId>
+      <artifactId>javax.servlet</artifactId>
+      <version>10.0-b28</version>
+    </dependency>
+
       <dependency>
         <groupId>org.glassfish.jersey.containers</groupId>
         <artifactId>jersey-container-servlet-core</artifactId>
@@ -43,7 +50,7 @@
 	  <dependency>
 	    <groupId>mysql</groupId>
 	    <artifactId>mysql-connector-java</artifactId>
-	    <version>8.0.22</version>
+	    <version>8.0.27</version>
 	</dependency>  
   </dependencies>
   <build>
