+++ EcoMap/src/main/java/com/GREENWORKS/controller/ShowPillars.java
@@ -1,5 +1,6 @@
 package com.GREENWORKS.controller;
 
+import com.GREENWORKS.DAO.*;
 import java.io.IOException;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -7,11 +8,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import com.GREENWORKS.DAO.*;
-
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 
 public class ShowPillars{
     private static final long serialVersionUID = 1L;
@@ -21,7 +17,7 @@ public class ShowPillars{
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {		
 		PillarDAO pd = new PillarDAO();
-		pd.showPillar();
+		pd.showPillars();
 
 		RequestDispatcher rd = request.getRequestDispatcher("index.jsp");
 		rd.forward(request, response);
