--- orlandoEcoMap_DataCollection/pom.xml
@@ -1,233 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>dataCollection</groupId>
-  <artifactId>maven</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-
-  <name>maven</name>
-  <!-- FIXME change it to the project's website -->
-  <url>http://www.example.com</url>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <maven.compiler.source>1.7</maven.compiler.source>
-    <maven.compiler.target>1.7</maven.compiler.target>
-    <arcgis.version>100.12.0</arcgis.version>
-  </properties>
-
-  <repositories>
-    <repository>
-      <id>arcgis</id>
-      <url>https://esri.jfrog.io/artifactory/arcgis</url>
-    </repository>
-  </repositories>
-
-  <dependencies>
-    <!--JavaFX dependencies-->
-    <dependency>
-      <groupId>org.openjfx</groupId>
-      <artifactId>javafx-controls</artifactId>
-      
-    </dependency>
-    <dependency>
-      <groupId>org.openjfx</groupId>
-      <artifactId>javafx-web</artifactId>
-      
-    </dependency>
-    <dependency>
-      <groupId>org.openjfx</groupId>
-      <artifactId>javafx-fxml</artifactId>
-      
-    </dependency>
-    <dependency>
-      <groupId>org.openjfx</groupId>
-      <artifactId>javafx-media</artifactId>
-      
-    </dependency>
-
-    <!--ArcGIS dependencies-->
-    <dependency>
-      <groupId>com.esri.arcgisruntime</groupId>
-      <artifactId>arcgis-java</artifactId>
-      <version>${arcgis.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>com.esri.arcgisruntime</groupId>
-      <artifactId>arcgis-java-jnilibs</artifactId>
-      <version>${arcgis.version}</version>
-      <type>zip</type>
-    </dependency>
-    <dependency>
-      <groupId>com.esri.arcgisruntime</groupId>
-      <artifactId>arcgis-java-resources</artifactId>
-      <version>${arcgis.version}</version>
-      <type>zip</type>
-    </dependency>
-
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
-    </dependency>
-    
-    <dependency>
-    <groupId>org.apache.poi</groupId>
-    <artifactId>poi-ooxml</artifactId>
-    <version>4.1.0</version>
-    </dependency>
-    
-    <dependency>
-    <groupId>mysql</groupId>
-    <artifactId>mysql-connector-java</artifactId>
-    <version>5.1.46</version>
-    <scope>runtime</scope>
-    </dependency>
-</dependencies>
-    
-  <build>
-    <pluginManagement><!-- lock down plugins versions to avoid using Maven defaults (may be moved to parent pom) -->
-      <plugins>
-        <!-- clean lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#clean_Lifecycle -->
-        <plugin>
-          <artifactId>maven-clean-plugin</artifactId>
-          <version>3.1.0</version>
-        </plugin>
-        <!-- default lifecycle, jar packaging: see https://maven.apache.org/ref/current/maven-core/default-bindings.html#Plugin_bindings_for_jar_packaging -->
-        <plugin>
-          <artifactId>maven-resources-plugin</artifactId>
-          <version>3.0.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>3.8.0</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-surefire-plugin</artifactId>
-          <version>2.22.1</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-jar-plugin</artifactId>
-          <version>3.0.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-install-plugin</artifactId>
-          <version>2.5.2</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-deploy-plugin</artifactId>
-          <version>2.8.2</version>
-        </plugin>
-        <!-- site lifecycle, see https://maven.apache.org/ref/current/maven-core/lifecycles.html#site_Lifecycle -->
-        <plugin>
-          <artifactId>maven-site-plugin</artifactId>
-          <version>3.7.1</version>
-        </plugin>
-        <plugin>
-          <artifactId>maven-project-info-reports-plugin</artifactId>
-          <version>3.0.0</version>
-        </plugin>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-dependency-plugin</artifactId>
-          <version>3.1.1</version>
-          <configuration>
-            <artifactItems>
-              <artifactItem>
-                <groupId>com.esri.arcgisruntime</groupId>
-                <artifactId>arcgis-java-jnilibs</artifactId>
-                <version>${arcgis.version}</version>
-                <type>zip</type>
-                <overWrite>false</overWrite>
-                <outputDirectory>${user.home}/.arcgis/${arcgis.version}</outputDirectory>
-              </artifactItem>
-              <artifactItem>
-                <groupId>com.esri.arcgisruntime</groupId>
-                <artifactId>arcgis-java-resources</artifactId>
-                <version>${arcgis.version}</version>
-                <type>zip</type>
-                <overWrite>false</overWrite>
-                <outputDirectory>${user.home}/.arcgis/${arcgis.version}</outputDirectory>
-              </artifactItem>
-            </artifactItems>
-          </configuration>
-        </plugin>
-        <plugin>
-          <groupId>org.codehaus.mojo</groupId>
-          <artifactId>exec-maven-plugin</artifactId>
-          <version>1.6.0</version>
-          <executions>
-            <execution>
-              <goals>
-                <goal>java</goal>
-              </goals>
-            </execution>
-          </executions>
-          <configuration>
-            <!--mainClass>com.example.app.App</mainClass--> <!--Change this to main class of the app-->
-          </configuration>
-        </plugin>
-        <plugin>
-          <groupId>io.takari</groupId>
-          <artifactId>maven</artifactId>
-          <version>0.7.4</version>
-        </plugin>
-      </plugins>
-    </pluginManagement>
-  </build>
-
-<!-- This section was created automatically upon unpacking and fixing dependencies.
-     Feel free to shift them upwards into the dependencies section if you don't like their placement. -->
-  <dependencyManagement>
-    <dependencies>
-
-      <dependency>
-        <groupId>commons-codec</groupId>
-        <artifactId>commons-codec</artifactId>
-        <version>1.12</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.openjfx</groupId>
-        <artifactId>javafx-base</artifactId>
-        <version>11.0.2</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.openjfx</groupId>
-        <artifactId>javafx-graphics</artifactId>
-        <version>11.0.2</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.openjfx</groupId>
-        <artifactId>javafx-controls</artifactId>
-        <version>11.0.2</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.openjfx</groupId>
-        <artifactId>javafx-media</artifactId>
-        <version>11.0.2</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.openjfx</groupId>
-        <artifactId>javafx-web</artifactId>
-        <version>11.0.2</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.openjfx</groupId>
-        <artifactId>javafx-fxml</artifactId>
-        <version>11.0.2</version>
-      </dependency>
-
-    </dependencies>
-  </dependencyManagement>
-</project>
