--- orlandoEcoMap_DataCollection/src/test/java/dataCellection/maven/AppTest.java
@@ -1,20 +1,0 @@
-package dataCellection.maven;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
